separate utils directory
