+++ learn/utils/aurora.py
@@ -1,6 +1,4 @@
 from learn.models import Technology, Domain, Resources
-import json
-import wikipedia
 
 def get_all_domains():
     """
@@ -117,46 +115,3 @@ def get_info_by_slug(slug):
         return_data = Technology.objects.get(name=name)
 
     return return_data
-
-def get_wiki_data(term):
-    """
-    Input: term, a string
-    Output: return_data, a dictionary
-    Returns page title, page url, page content for term
-    Returns None if it encounters a ConnectionError
-    """
-    return_data = {}
-    try:
-        term_data = wikipedia.page(term)
-        return_data['title'] = term_data.title
-        return_data['url'] = term_data.url
-        return_data['content'] = term_data.content
-    except ConnectionError:
-        print "Please check your internet connection"
-        return None
-
-def get_similar_search(term):
-    """
-    Input: term, a string
-    Output: return_data, a list
-    Returns a list containing strings which are similar to the current search term
-    Returns None if it encounters a ConnectionError
-    """
-    try:
-        return_data = wikipedia.search(term)
-    except ConnectionError:
-        print "Please check your internet connection"
-        return None
-
-def get_wiki_summary(term, sentences=3):
-    """
-    Input: term, a string. sentences, an interger
-    Output: string
-    Returns the summary for term with a constraint on the number of sentences. Default = 3
-    Returns None if it encounters a ConnectionError
-    """
-    try:
-        return_data = wikipedia.summary(term, sentences)
-    except ConnectionError:
-        print "Please check your internet connection"
-        return None
